+++ src/search-filter-lambda/src/search-filter-lambda-percolate-product/src/service.rs
@@ -512,6 +512,8 @@ mod tests {
             updated_by: Actor::System,
             created: OffsetDateTime::now_utc(),
             updated: OffsetDateTime::now_utc(),
+            last_hybrid_search_matched: OffsetDateTime::now_utc(),
+            embedding: None,
         }
     }
 
@@ -534,6 +536,8 @@ mod tests {
             updated_by: Actor::System,
             created: OffsetDateTime::now_utc(),
             updated: OffsetDateTime::now_utc(),
+            last_hybrid_search_matched: OffsetDateTime::now_utc(),
+            embedding: None,
         }
     }
 
@@ -817,6 +821,8 @@ mod tests {
             updated_by: Actor::System,
             created: OffsetDateTime::now_utc(),
             updated: OffsetDateTime::now_utc(),
+            last_hybrid_search_matched: OffsetDateTime::now_utc(),
+            embedding: None,
         };
 
         let mut get_service = MockGetProductService::default();