+++ src/search-filter/src/core/user_search_filter_search.rs
@@ -3,4 +3,5 @@ use common::resource_state::domain::ResourceState;
 #[derive(Debug, Clone, PartialEq, Eq, Default)]
 pub struct UserSearchFilterSearch {
     pub state: Option<ResourceState>,
+    pub has_enhanced_search_description: Option<bool>,
 }