+++ src/user-lambda/src/user-lambda-tier-update/src/lib.rs
@@ -180,6 +180,7 @@ fn search_filter_state_update(state: ResourceStateRecord) -> UserSearchFilterRec
         notifications: None,
         state: Some(state),
         enhanced_search_description: None,
+        embedding: None,
         product_query: None,
         shop_name_query: None,
         exclude_shop_name_query: None,
@@ -203,6 +204,7 @@ fn search_filter_state_update(state: ResourceStateRecord) -> UserSearchFilterRec
         currency: None,
         updated_by: ActorRecord::System,
         updated: OffsetDateTime::now_utc(),
+        last_hybrid_search_matched: None,
     }
 }
 
@@ -248,6 +250,8 @@ mod tests {
             updated_by: Actor::User(user.user_id),
             created: OffsetDateTime::now_utc(),
             updated: OffsetDateTime::now_utc(),
+            last_hybrid_search_matched: OffsetDateTime::now_utc(),
+            embedding: None,
         };
         filter.into()
     }